requirements.txt
python-dotenv
numpy
pandas
setuptools
certifi
pymongo
pymongo[srv]==3.12
scikit-learn
#-e .
